// src/types/news.ts
export interface Article {

    id: number;

    title: string;

    description: string;

    content: string;

    image: string;

    image_url: string;

    source: string;

    category: string;

    country: string;

    ai_tags: string[];

    published_at: string;

    summary: string;

    url: string;

    author: string;

    language: string;

    sentiment: string;

    reading_time: number;

}